crawler/src/requirements.txt
pandas~=1.4.0
mysql-connector-python~=8.0.28
beautifulsoup4~=4.10.0
lxml~=4.9.2